Final Application V1.2
This version is adding data to the table on the Exercise Selection view controller, when you select an exercise and hit the plus button it sends that to the initial view controller adding it to a table. This is stored in a dictionary along with the current date. I have attempted to store this dictionary in a plist which is seemingly fine, but when I try to load the data I get an error crash. 'Cannot cast value of type __NSCFConstantString to NSDictionary'